Desacomple funciones en un modulo de utils y agregue nuevo parametro de tipo de proyecto
