+++ utils/dirFiles.js
@@ -1,0 +1,71 @@
+const fs = require('fs');
+const path = require('path');
+const util = require('util');
+
+var mkdir = function(dir) {
+	// making directory without exception if exists
+	try {
+		fs.mkdirSync(dir, 0755);
+	} catch(e) {
+        rmdir(NEW_PATH);
+		if(e.code != "EEXIST") {
+            
+			throw e;
+		}
+	}
+};
+
+var rmdir = function(dir) {
+	if (path.existsSync(dir)) {
+		var list = fs.readdirSync(dir);
+		for(var i = 0; i < list.length; i++) {
+			var filename = path.join(dir, list[i]);
+			var stat = fs.statSync(filename);
+			
+			if(filename == "." || filename == "..") {
+				// pass these files
+			} else if(stat.isDirectory()) {
+				// rmdir recursively
+				rmdir(filename);
+			} else {
+				// rm fiilename
+				fs.unlinkSync(filename);
+			}
+		}
+		fs.rmdirSync(dir);
+	} else {
+		console.warn("warn: " + dir + " not exists");
+	}
+};
+
+
+var copy = function(src, dest) {
+    console.log(`* Copiando ${src} a ${dest}-------`);
+	var oldFile = fs.createReadStream(src);
+	var newFile = fs.createWriteStream(dest);
+    // util.pump(oldFile, newFile);
+    oldFile.pipe(newFile);
+    oldFile.on(`end`,(e)=>{
+        newFile.close();
+        oldFile.close();
+        console.log(`copiado...`)
+    })
+};
+
+var copyDir = function(src, dest) {
+	mkdir(dest);
+	var files = fs.readdirSync(src);
+	for(var i = 0; i < files.length; i++) {
+		var current = fs.lstatSync(path.join(src, files[i]));
+		if(current.isDirectory()) {
+			copyDir(path.join(src, files[i]), path.join(dest, files[i]));
+		} else if(current.isSymbolicLink()) {
+			var symlink = fs.readlinkSync(path.join(src, files[i]));
+			fs.symlinkSync(symlink, path.join(dest, files[i]));
+		} else {
+			copy(path.join(src, files[i]), path.join(dest, files[i]));
+		}
+	}
+};
+
+module.exports = copyDir;
